+++ packages/react-native/src/index.ts
@@ -1,11 +1,18 @@
 // Export components
 export * from "./components/Activity/ActivityItem";
 export * from "./components/Button";
-export * from "./components/EmojiAvatar";
 export * from "./components/ExampleComponent";
 export * from "./components/Icon";
-export * from "./components/IconAvatar";
-export * from "./components/Information/ModuleAvatar";
+export * from "./components/Avatars/exports";
+export * from "./components/Counter";
+export * from "./components/OnePreset";
+export * from "./components/Tags/exports";
+export * from "./components/OneChip";
+export * from "./components/Navigation/PageHeader";
+export * from "./components/Tags/exports";
 
 // Export icons
 export * from "./icons";
+
+// Export modules
+export * from "./modules/inbox/InboxView";