+++ packages/react/src/experimental/Lists/DataList/index.tsx
@@ -1,5 +1,7 @@
 import { forwardRef, ReactElement } from "react"
 
+import { DotTag, DotTagProps } from "@/experimental/Information/Tags/DotTag"
+import { cn } from "@/lib/utils"
 import { IconType } from "../../../components/Utilities/Icon"
 import { CompanyAvatar } from "../../Information/Avatars/CompanyAvatar"
 import { PersonAvatar } from "../../Information/Avatars/PersonAvatar"
@@ -9,6 +11,7 @@ import { InternalActionType, ItemContainer } from "./ItemContainer"
 export type DataListProps = {
   children: ReactElement<Items>[] | ReactElement<Items>
   label?: string
+  isHorizontal?: boolean
 }
 
 type Items =
@@ -18,13 +21,26 @@ type Items =
   | typeof TeamItem
 
 const _DataList = forwardRef<HTMLUListElement, DataListProps>(
-  ({ children, label }, ref) => {
+  ({ children, label, isHorizontal = false }, ref) => {
     return (
-      <div className="min-w-32 max-w-72">
-        {label && (
-          <p className="mb-0.5 px-1.5 text-f1-foreground-secondary">{label}</p>
+      <div
+        className={cn(
+          isHorizontal
+            ? "flex min-h-12 flex-1 flex-col py-1.5 pl-3 pr-1.5 xs:flex-row"
+            : "min-w-32 max-w-72"
         )}
-        <ul className="flex flex-col gap-0.5" ref={ref}>
+      >
+        {label && (
+          <p
+            className={cn(
+              "px-1.5 text-f1-foreground-secondary",
+              isHorizontal ? "mt-1.5 w-44 xs:px-0" : "mb-0.5"
+            )}
+          >
+            {label}
+          </p>
+        )}
+        <ul className="flex flex-col justify-center gap-0.5" ref={ref}>
           {children}
         </ul>
       </div>
@@ -141,6 +157,20 @@ const TeamItem = forwardRef<HTMLLIElement, TeamItemProps>(
 
 TeamItem.displayName = "TeamItem"
 
+type DotTagItemProps = DotTagProps
+
+const DotTagItem = forwardRef<HTMLLIElement, DotTagItemProps>(
+  ({ ...props }, ref) => {
+    return (
+      <li ref={ref} className="flex items-start pt-1">
+        <DotTag {...props} />
+      </li>
+    )
+  }
+)
+
+DotTagItem.displayName = "DotTagItem"
+
 /**
  * convert simplified action type received from user to internal action format
  * @param action ActionType
@@ -162,4 +192,5 @@ export const DataList = Object.assign(_DataList, {
   CompanyItem,
   PersonItem,
   TeamItem,
+  DotTagItem,
 })