+++ packages/react/src/experimental/RichText/exports.tsx
@@ -1,3 +1,5 @@
+export * from "./BasicTextEditor"
+export * from "./CoreEditor/types"
 export * from "./FileItem"
 export * from "./RichTextDisplay"
 export * from "./RichTextEditor"